抽取异常类
