+++ src/main/kotlin/net/slardar/slardarhttp/exception/SlardarHTTPException.kt
@@ -1,0 +1,7 @@
+package net.slardar.slardarhttp.exception
+
+class SlardarHTTPException(message: String, private val errorCode: Int) : Exception(message) {
+    fun getErrorCode(): Int {
+        return errorCode
+    }
+}
