+++ src/main/kotlin/net/slardar/slardarhttp/exception/SlardarHTTPJSONException.kt
@@ -1,0 +1,11 @@
+package net.slardar.slardarhttp.exception
+
+import org.json.JSONObject
+
+@Suppress("MemberVisibilityCanBePrivate", "CanBeParameter")
+class SlardarHTTPJSONException(val jsonMessage: JSONObject, private val errorCode: Int) :
+    Exception(jsonMessage.toString()) {
+    fun getErrorCode(): Int {
+        return errorCode
+    }
+}
